updated jest testing to include all service method
