+++ lib/interfaces/index.ts
@@ -1,1 +1,2 @@
 export * from './sentry-options.interface';
+export * from './sentry-service.interface';
